+++ Reach/test/utopia/reach/test/ReachTestContext.scala
@@ -65,6 +65,7 @@ object ReachTestContext
 	// INIITAL CODE -----------------------
 	
 	KeyboardEvents.specifyExecutionContext(exc)
+	KeyboardEvents.setupKeyDownEvents(actorHandler)
 	
 	
 	// OTHER    --------------------------