+++ Reach/Changes.md
@@ -1,13 +1,14 @@
 # Utopia Reach - List of Changes
 
 ## v1.3.1 (in development)
-Built with Scala v2.13.14
-Working on slider
 ### New features
+- Added a **Slider** component
 - Added **HasGuiState** trait
 ### New methods
 - **PartOfComponentHierarchy**
   - Added `.linkedFlag`
+### Other changes
+- Built with Scala v2.13.14
 
 ## v1.3 - 28.07.2024
 This update follows changes introduced in **Genesis v4.0**, which mostly affect the component